ml-engine/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
shapely==2.0.4
haversine==2.8.1
python-dotenv==1.0.1

# Ethereum Blockchain
web3>=6.0.0
eth-account>=0.10.0
py-solc-x>=2.0.0

